Add scaleway provider to the examples
I wanted to try this quickstart at Scaleway and it was not one of the
examples. There is one PR open for this but it was WIP and using an
outdated version of scaleway/scaleway terraform provider.

https://github.com/rancher/quickstart/pull/123

I based it in the DigitalOcean example because they clouds and terraform
providers are very similar.
What changed to help with review:

- region/zone by provider instead of per resource/instance
- `public_ip` attributes instead of `ipv4`
- ssh key/credentials are added to project instead of each instance
- image names and instance size/type

Also changed the README and the `terraform.yml` github workflow.

Tested it and the rancher server and a RKE example workload cluster
started properly and everything seems to be working as expected.

- Scaleway instances - https://i.imgur.com/po81A5d.png
- Rancher clusters global view - https://i.imgur.com/U4GFHHM.png